test: extract base modbus transport tests
